==> codes/1.Feature_Extraction.ipynb <==
## Feature extraction : to vectorize given words 

* we are converting each word to a vectorized form which is a binary vector ( containing only 0s and 1s )
* Each feature/column is a binary question. e.g. is first letter capitalized or not ? ( Yes = 1 / No = 0 ) 
* We are considering mainly suffix , prefix , prev and next context , some english rules of words as features
* At the end we will store this feature dataframe to a csv file , for future use : as this code takes time

# importing necessary packages

import pandas as pd
import numpy as np
import time

### Loading training data

df = pd.read_csv('./Data_for_postagging/postrain', sep = " ")
df.head()
df.shape

# final dataframe : it ll be our data matrix 

final= pd.DataFrame(index = range(211730) , columns = ['word', 'tag', 'capitalize', 
                                                       'digit_first', 
                                                       's_dig__e_alpha', 'p_anti', 
                                                       'p_pre', 'p_un',
                                                       'p_dis', 'p_inter', 'p_mis', 
                                                       'p_non', 'p_over_under',
                                                       'p_in_im', 'p_en_em','s_able',
                                                       's_al_ial', 
                                                      's_ed_ing','s_tion_ion','s_est',
                                                      's_less', 's_e_es', 
                                                      's_en','s_ly','s_er','s_\'s_s\'',
                                                      'prev_word'])

#prev_word represents that prev word is a, an or the
final

#df = pd.DataFrame(index=range(numRows),columns=range(numCols))

#final['word'][2] = 1
final.shape
final.head()

### Extracting feautures 

for i in df.index.values:
    
    j += 1
    
   # if(j > 10):
   #     break
    
    if df.word[i][0].isupper():
        final['capitalize'][i] = 1
    else:
        final['capitalize'][i] = 0
        
    if df.word[i][0].isdigit():
        final['digit_first'][i] = 1
    else:
        final['digit_first'][i] = 0
        
    if df.word[i][0].isdigit() and df.word[i][-1].isalpha():
        final['s_dig__e_alpha'][i] = 1
    else:
        final['s_dig__e_alpha'][i] = 0
        
    if df.word[i].startswith('anti'):
        final['p_anti'][i] = 1
    else:
        final['p_anti'][i] = 0
        
    if df.word[i].startswith('pre'):
        final['p_pre'][i] = 1
    else:
        final['p_pre'][i] = 0
        
    if df.word[i].startswith('un'):
        final['p_un'][i] = 1
    else:
        final['p_un'][i] = 0
    if df.word[i].startswith('dis'):
        final['p_dis'][i] = 1
    else:
        final['p_dis'][i] = 1
    if df.word[i].startswith('inter'):
        final['p_inter'][i] = 1
    else:
        final['p_inter'][i] = 0
    if df.word[i].startswith('mis'):
        final['p_mis'][i] = 1
    else:
        final['p_mis'][i] = 0
    if df.word[i].startswith('non'):
        final['p_non'][i] = 1
    else:
        final['p_non'][i] = 0
        
    if df.word[i].startswith('over') or df.word[i].startswith('under'):
        final['p_over_under'][i] = 1
    else:
        final['p_over_under'][i] = 1
        
    if df.word[i].startswith('in') or df.word[i].startswith('im'):
        final['p_in_im'][i] = 1
    else:
        final['p_in_im'][i] = 0
    if df.word[i].startswith('en') or df.word[i].startswith('em'):
        final['p_en_em'][i] = 1
    else:
        final['p_en_em'][i] = 0
        
    if df.word[i].endswith('able'):
        final['s_able'][i] = 1
    else:
        final['s_able'][i] = 0
        
    if df.word[i].endswith('al') or df.word[i].endswith('ial'):
        final['s_al_ial'][i] = 1
    else:
        final['s_al_ial'][i] = 0
        
    if df.word[i].endswith('ed') or df.word[i].endswith('ing'):
        final['s_ed_ing'][i] = 1
    else:
        final['s_ed_ing'][i] = 0
        
    if df.word[i].endswith('tion') or df.word[i].endswith('ion'):
        final['s_tion_ion'][i] = 1
    else:
        final['s_tion_ion'][i] = 0
    if df.word[i].endswith('est'):
        final['s_est'][i] = 1
    else:
        final['s_est'][i] = 0
    if df.word[i].endswith('less'):
        final['s_less'][i] = 1
    else:
        final['s_less'][i] = 0
    if df.word[i].endswith('e') or df.word[i].endswith('es'):
        final['s_e_es'][i] = 1
    else :
        final['s_e_es'][i] = 0
    if df.word[i].endswith('en'):
        final['s_en'][i] = 1
    else:
        final['s_en'][i] = 0
    if df.word[i].endswith('ly'):
        final['s_ly'][i]=1
    else:
        final['s_ly'][i]=0
    if df.word[i].endswith('er'):
        final['s_er'][i] = 1 
    else:
        final['s_er'][i] = 0
    if df.word[i].endswith('\'s') or  df.word[i].endswith('s\''):
        final['s_\'s_s\''][i] = 1
    else:
        final['s_\'s_s\''][i] = 0
    
    if i >= 1:
        if df.word[i-1] == 'a' or df.word[i-1] == 'an' or df.word[i-1] == 'the':
            final['prev_word'][i] = 1
        else:
            final['prev_word'][i] = 0
    else:
        final['prev_word'][i] = 0
        
    final['word'][i]= df.word[i]
    final['tag'][i]= df['pos_tag'][i]
    
    #if(j % 20000 == 0):
    #    print(j)

# append tks 2 mins for first 20000 and then time increases exponentially
# fixed size array tks 10 mins for every 20k rows

# NEVER EVER use append (means arraylist) when u knw array size in advance 
# performance of array is way way better than arraylist 

### Storing featurized dataframe to disk : for future use 

# storing the featurized df to disk into csv format 
final.to_csv('./final_df_by_arr.csv',sep = ',' , index = False , encoding = 'utf-8')

# reading from disk 
checkdf = pd.read_csv('./final_df.csv' , sep = ',' )
print(checkdf.shape)
checkdf.head()

checkdf['tag'].value_counts()
